Update the LDAP SDK to make some changes around the way that it treats attribute type definitions whose SYNTAX element contains an OID with the optional minimum upper bound portion present.
Previously, the AttributeTypeDefinition.getSyntaxOID methods always returned
the syntax OID exactly as it was provided in the attribute type definition,
which may or may not have included a minimum upper bound component.  This is
still the case, but I have added some new methods to make it possible to
identify the numeric OID and minimum upper bound portions separately.  In
particular, the getBaseSyntaxOID methods will extract only the numeric OID
portion (and will behave identically to getSyntaxOID if there is no minimum
upper bound portion), and the getSyntaxMinimumUpperBound method will return
the minimum upper bound portion, or -1 if no minimum upper bound is defined.

In addition, the Schema.getAttributeSyntax method has been updated so that it
will ignore any minimum upper bound portion that may be included in the
provided OID.  If the OID provided is only numeric, then it will work as
before without any changes.  If the provided OID also contains a minimum upper
bound in curly braces, then that portion will be ignored and only the numeric
portion of the OID will be used to retrieve the corresponding syntax definition
(whereas before it would have always returned null).