Update the LDAP SDK to provide a much more useful exception in the event that the connection is terminated as a result while trying to read a response from the server.
As an example, this used to be the exception thrown if the server sent a
malformed search result done message:

     An error occurred while processing the search:  The connection to server
     localhost:389 was closed while waiting for a response to search request
     SearchRequest(baseDN='', scope=BASE, deref=NEVER, sizeLimit=0,
     timeLimit=0, filter='(objectClass=*)', attrs={}).

This message doesn't really provide much information about whether the
connection was closed by the client or the server, and what the underlying
reason might have been.  The new message that will be generated in this case
is:

     An error occurred while processing the search:  The connection to server
     localhost:389 was closed while waiting for a response to search request
     SearchRequest(baseDN='', scope=BASE, deref=NEVER, sizeLimit=0,
     timeLimit=0, filter='(objectClass=*)', attrs={}):  Unable to read or
     decode an LDAP result:  The ASN.1 stream reader has already read beyond
     the end of this sequence (expected sequence of length 7 to end at 28 bytes
     into the stream, but 33 bytes have already been read from the stream).

At the very least, this message makes it clearer that there was a problem with
what the server sent to the client, and the additional information may help
someone figure out exactly where the problem is (if nothing else, then perhaps
by looking at a network packet capture).

In addition, the result code for the LDAPException may more accurately reflect
the condition that caused this problem.  In the past, it would always be
SERVER_DOWN, but it now may be something more appropriate.  For example, in the
above case it would be DECODING_ERROR.