Update the LDAP SDK so that the WakeableSleeper class uses Object.wait to perform the sleep rather than ArrayBlockingQueue.poll, and it uses Object.notifyAll to trigger the wakeup rather than ArrayBlockingQueue.offer. All of the documented semantics of the class (including the fact that any WakeableSleeper instance may only be used to cause one thread to sleep at any given time) will be preserved.
This change is necessary to work around an apparent bug in the Java platform in
which sometimes ArrayBlockingQueue.poll does not return after the specified
length of time (this also seems to happen when System.nanoTime consistently
starts returning the same value for every call regardless of how much time has
passed).  If the JVM enters this state, then the previous implementation of
WakeableSleeper.sleep would never return until WakeableSleeper.wakup was
called.