Update the LDAP SDK persistence framework to add a lazilyLoad element to the LDAPField annotation.  This can be used to indicate that the associated attribute should not be requested by default when retrieving an entry.  Also, add a new LDAPPersister.lazilyLoad method that can be used to load all or a subset of those fields marked for lazy loading.
This can be useful for cases in which an attribute is not always needed and can
be expensive to retrieve, for example because it is very large or requires
expensive server-side processing.  This makes it possible to pay the cost of
retrieving the expensive attribute(s) only when they are needed.