Update the LDAP SDK to make the LDAPConnection.reconnect() method public, which will allow it to be used by external clients.  This method can be used to request that a connection be closed (if necessary) and re-established to the server to whit was last connected.  If appropriate, an attempt will also be made to re-authenticate that connection.
In addition, the SearchRate, ModRate, and AuthRate tools have been updated to
use this reconnect method to better handle server unavailability while they are
running.  Previously, if the server became unavailable then they would make
one attempt to reconnect and if that failed then each subsequent search attempt
would result in a "server down" result, even if the server later came back up.
With this commit, the client will attempt to re-establish the connection after
each failure which indicates that the connection may no longer be usable.  If
the server does come back up, then the tool will be able to automatically
reconnect and resume placing load on it.