This commit addresses a problem with how example command line arguments are quoted.  The bulk of this commit introduces a new class, ExampleCommandLineArgument, that handles the argument quoting.
This class provides access to a form of a command-line argument that is
safe to use in a shell.  It includes both forms for both Unix (bash shell
specifically) and Windows, since there are differences between the two
platforms.  Quoting of arguments is performed with the following goals:

  * The same form should be used for both Unix and Windows whenever
    possible.

  * If the same form cannot be used for both platforms, then make it
    as easy as possible to convert the form to the other platform.

  * If neither platform requires quoting of an argument, then it is not
    quoted.

To that end, here is the approach that we've taken:

  * Characters in the output are never escaped with the \ character
    because Windows does not understand \ used to escape.
  
  * On Unix, double-quotes are used to quote whenever possible since
    Windows does not treat single quotes specially.

  * If a String needs to be quoted on either platform, then it is quoted
    on both.  If it needs to be quoted with single-quotes on Unix, then
    it will be quoted with double quotes on Windows.

  * On Unix, single-quote presents a problem if it's included in a
    string that needs to be singled-quoted, for instance one that includes
    the $ or ! characters.  In this case, we have to wrap it in 
    double-quotes outside of the single-quotes.  For instance, Server's!
    would end up as 'Server'"'"'s!'.

  * On Windows, double-quotes present a problem.  They have to be
    escaped using two double-quotes inside of a double-quoted string.
    For instance "Quoted" ends up as """Quoted""".

All of the forms can be unambiguously parsed using the
parseExampleCommandLine method regardless of the platform.  This
method can be used when needing to parse a command line that was generated
by this class outside of a shell environment, e.g. if the full command line
was read from a file.  Special characters that are escaped include |, &, ;,
(, ), !, ", ', *, ?, $, and `.

The StaticUtils#cleanExampleCommandLineArgument method, which
previously provided this functionality, has been deprecated with
a reference to the new ExampleCommandLineArgument class.