Update the LDAP SDK so that attempts to process an extended operation is more likely to throw an exception when there is a legitimate failure.  Because extended operations often return non-success results under normal conditions (e.g., the cancel extended operation will never return success), the LDAP SDK will not necessarilly throw an exception simply because it gets a non-success result.  However, this has caused problems in some cases and it is better to throw an exception in cases where there is obviously a problem (e.g., the result indicates the connection is no longer usuable or the server may be in a weird state).  As a result of this change, the LDAP SDK is now more likely to throw an exception under conditions where it is confident that some kind of problem occurred during processing, although simply because it didn't throw an exception still does not mean that the requested operation was necessarily processed successfully.
As a side effect, this improves the behavior of the LDAP SDK when attempting
to process extended operations using connections from a connection pool.
Previously, the connection pool did not examine the result from an extended
operation that didn't throw an exception, and under some conditions it could
continue trying to use a connection when it should have been able to determine
that it was no longer valid.