Update the LDAP SDK in-memory directory server to provide the ability to create and restore snapshots of the data in the server.  Snapshots include both user data and changelog records, but not server configuration.
This can be used to create a base data set that you can roll back to in order
to have a clean state for running tests.  It may also be useful in adding
future functionality (e.g., LDAP transactions) in which it may be necessary
to roll back already-applied changes if a failure is encountered for a
subsequent change in the same set.