Update the LDAP SDK to add a new PasswordProvider API which can be used to obtain a password rather than using a hard-coded value.  This is a security feature, and the array containing the password, as well as buffers used in the course of generating the encoded request which also contains the password, will be zeroed out after the content is no longer needed in order to avoid keeping the clear-text password in memory.
At present, this is only available for use with simple authentication.
However, it may be extended to other areas in which clear-text passwords may
be needed (including the CRAM-MD5, DIGEST-MD5, GSSAPI, and PLAIN SASL
mechanisms, and the password modify extended request).

A ReadFromFilePasswordProvider implementation has been provided which can be
used to read a password from a file.  In addition, the PasswordProvider API is
public, so third-party developers can create custom implementations if
desired.