Update the LDAP SDK to fix a bug in which the configured referral connector could have been ignored when establishing a connection in order to follow referrals.
Also, update the LDAPConnectionOptions object to make it possible to configure
the referral connector to be used for associated connections rather than
requiring the referral connector to be explicitly set on a given connection.
It is still possible to set the connector on a per-connection basis, and if
that is done then it will override the connector set in the connection
options.