Update the in-memory directory server to add the ability to control which types of operations will be allowed, and which types of operations will only be allowed for authenticated clients.  By default, all operations will be allowed, and no operations will require authentication.
Note, however, that even if a given type of operation is not allowed for LDAP
clients, then those operations may still be processed by calling the
appropriate methods in the InMemoryDirectoryServer class.  Similarly,
operations processed through InMemoryDirectoryServer methods will be
considered authenticated, so they will be allowed regardless of whether
authentication is required for those operations.