Update the LDAP SDK to provide a new DNSSrvRecordServerSet class, whcih is a ServerSet implementation that can use DNS to discover available directory servers using SRV records, as described in RFC 2782.  SRV records include a number of elements, including:
* The address of the server
* The port of the server
* The priority of the server relative to other servers (servers with a lower
  priority will be tried before servers with a higher priority)
* The weight of the server relative to other servers with the same priority
 (if multiple servers have the same priority, then the weights will be used to
 indicate how frequently each should be used relative to the others)

The new server set implementation can retrieve these SRV records and use them
to establish connections.  If multiple servers are defined for the specified
service, then it will use the priority and weight of each server to determine
the appropriate order in which to try to contact them to get a connection.