Update the LDAP SDK to add a connection option that controls whether to allow multiple threads can use the same socket factory instance to create sockets concurrently.
A recent change caused all socket factory interaction to be synchronized so
that it was only possible to create a single socket at any one time, since
some IBM VMs were observed to misbehave when trying to create multiple SSL
sockets concurrently from the same socket factory.  However, this introduced
an undesirable side effect that could interfere with the ability to obtain a
connection in a reasonable length of time in the event that one of the target
systems was unavailable in a way that made it completely unresponsible (e.g.,
if the system was powered off or there was a network failure).  The new
default behavior is to allow concurrent socket factory use on Sun/Oracle JVMs,
but require synchronized use on JVMs from all other vendors.  The
LDAPConnectionOptions class offers methods that can be used to override this
behavior if desirable.