Revise the previous commit that introduced a SynchronizedSocketFactory and the ability to configure whether concurrent socket factory use should be allowed. A new method has been added which makes it possible to obtain the wrapped socket factory so that it is available for inspection if necessary.
Also, create a new SynchronizedSSLSocketFactory class that is itself an
SSLSocketFactory and will be used to provide a synchronized wrapper for
SSLSocketFactory instances.  This can be useful for cases in which a caller
retrieves the socket factory associated with a connection and checks to see if
it is an instance of SSLSocketFactory in order to make a determination about
whether the associated connection was created as a secure connection.