Update the LDAP SDK to add a usePooledSchema method as an alternative to useSchema in the LDAPConnectionOptions class.  This new option should only be set to true when the connection options are to be used with a connection pool, and if that is the case then connections in that pool will all share the same schema object.  This can help improve performance and reduce network consumption over the useSchema option, since the schema will information will only be retrieved once per pool rather than once per connection.
Note that it is also possible to associate a timeout with the pooled schema,
and by default that timeout is set to one hour.  If the pooled schema is older
than that timeout, then the next attempt to establish a new connection will
cause the schema to be re-retrieved from the server, and that schema will be
used for new connections until it expires.