Update the LDAP SDK so that the LDIF representation of modify change records will include a trailing dash after the last change, in accordance with RFC 2849.  Previously, the LDIF representation would not include a trailing dash, which is a popular convention but incompatible with the strict interpretation of the LDIF specification.
It is still possible to configure the LDAP SDK to provide the old behavior if
that is preferable for some reason (e.g., if an application that will consume
the resulting LDIF content cannot accept the trailing dash).  There are static
methods that can be used to indicate that the old behavior should be
maintained, and a system property has been added that can be used to retain
the former behavior without any code changes.