Update the LDAP SDK so that the implementations of the authrate, modrate, search-and-mod-rate, and searchrate tools include support for varying the target rate over time in accordance with a definition contained in a text file. The path to this file is specified via the new "--variableRateData" argument.
The file must begin with a header in which each line begins with the "#"
character and has a key-value pair separated by an equal sign.  The supported
header keys include:

- format -- This specifies the format for the variable rate data.  This header
  must be provided.  At present, the only supported format is "rate-duration".

- repeat -- This indicates whether the tool should start processing the
  variable rate data from the beginning again once the end of the data has been
  reached (if the value is "true"), or if the tool should exit (if the value is
  "false").  This header is optional, and if it is not provided then the tool
  will exit after the end of the variable rate data has been reached.

- default-duration -- This specifies the default duration that should be used
  for any rate definition line that does not specify a duration.  This header
  is optional, and if it is not provided then all rate definition lines must
  include a duration.

The remainder of the file should include rate definition lines that contain a
target rate followed by a comma and the length of time that rate should be
maintained.  The target rate may be either an integer value to specify the
absolute rate per second, or it may be a floating-point value followed by the
letter "x" to indicate that the floating-point value should be used as a
multiplier in conjunction with the value specified by the "--ratePerSecond"
argument.  The duration should be specified as an integer value followed by a
time unit that may be one of the following:

- nanoseconds, nanosecond, nanos, nano, ns
- microseconds, microsecond, micros, micro, us
- milliseconds, millisecond, millis, milli, ms
- seconds, second, secs, sec, s
- minutes, minute, mins, min, m
- hours, hour, hrs, hr, h
- days, day, d

If a default-duration header is provided, then the duration may be omitted from
one or all of the rate definition lines if the default duration should be used
for that line.

Any blank lines in the file will be ignored, as will any non-header lines
beginning with the "#" character.  The "#" character may also be used in the
middle of a line, and anything after that character will be ignored.

The following example may be used as a variable rate data file that will cause
the rate to increase from 1,000 operations per second to 10,000 operations per
second in increments of 1,000 operations per second over a ten-minute period.
Once the rate has decreased back down to 1,000 operations per second, it will
start over and begin climbing back up to 10,000:

     # format=rate-duration
     # repeat=true
     1000, 1 minute
     2000, 1 minute
     3000, 1 minute
     4000, 1 minute
     5000, 1 minute
     6000, 1 minute
     7000, 1 minute
     8000, 1 minute
     9000, 1 minute
     10000, 1 minute