Update the LDAP SDK Entry.diff method to account for differences in attribute options in the set of attributes to examine versus attributes actually contained in the provided entries.
Previously, if the diff method was given a specific set of attributes to
examine, then it would only perform a case-insensitive comparison of the
provided attribute names with the names of the attributes included in the
source and target entries.  For example, if the attributes to examine
included "userCertificate" then it would incorrectly ignore an attribute named
"userCertificate;binary".

With this change, if the diff method is provided with a set of attributes to
examine, then it will match the base names (without any options) of the
attributes in that list against the base names of the attributes contained in
the source and target entries.