Update the LDAP SDK's JSON implementation to add support for comments in the string representation of JSON objects.  Although the JSON specification technically does not allow for comments, the LDAP SDK JSON decoder will allow comments that are surrounded by /* and */, or comments that start with // and continue until the end of the line.
This can be useful for cases in which a JSON object is used for configuration
purposes and that configuration includes comments to describe what is
permitted or to exclude elements without removing them from the object.  The
JSON parser will ignore all comments, so there is no way to access them
programmatically, nor is there any way to include comments when
programmatically creating a JSON object.