Update the LDAP SDK argument parser to add support for an interactive mode. This mode makes it possible for a tool to interactively prompt the user for the LDAP connection and authentication arguments, as well as any required tool-specific arguments.  It will then display a menu that allows the user to edit the values for the tool arguments, run the tool with the selected arguments, and display the command line that may be used to run the tool with those arguments in non-interactive mode.
Most of the example tools have been updated to support interactive mode, and
to default to starting in interactive mode if the tool is invoked without any
arguments.

Some additional argument-related cleanup has been performed, and improvements
have been made to the shell scripts and batch files used to launch the example
tools.