Update the LDAP SDK argument parser to add support for specifying argument values using a properties file.  The path to the properties file can be explicitly specified using an argument, or it can be specified using a Java property or an environment variable.  If a properties file is available, it can be used to obtain default values for any arguments not provided on the command line.
The names of the properties in the file should match the name of the
corresponding arguments.  If the argument should apply only to one particular
tool, then the property name should start with a tool name and a period.  For
example, the "hostname" property can be used to specify a default value for
the "hostname" property for any tool that supports it, while the
"ldapsearch.hostname" property can be used to specify a default hostname that
will only be used by the ldapsearch tool.  If both tool-specific and generic
versions of a property are defined, then the tool-specific version will take
precedence.

A properties file can include information about properties that don't apply to
the tool that is being run.  This makes it possible to have a single
properties file that can be shared by multiple tools if desired.

A new "--generatePropertiesFile" argument is also available.  That argument
can be used to generate a properties file with detailed comments about all
of the supported arguments for that tool.