Update the LDAP SDK to add an entry transformation that can be used to flatten a DIT below a specified base DN, so that all entries below that base DN will be exactly one level below that base DN.  This can be used to convert a highly-branched LDIF to one in which all of the entries are below a common parent.  Any name-value pairs that are stripped out of the DN during the flattening process can optionally be added into the entry as attribute values, and it is also possible to optionally convert the original RDN into a multivalued RDN that contains all name-value pairs from the original RDN plus any components that between that RDN and the flatten base DN.
The transformation also provides an option to define an exclude filter, and if
an exclude filter is provided the resulting LDIF will exclude any entry that
resides below the flatten base DN that matches the given filter.  This can be
used to eliminate "container" entries whose only purpose in the original LDIF
had been to provide hierarchy that is being eliminated.

Finally, the transform-ldif tool has been updated to provide arguments that
allow it to take advantage of this new flatten subtree transformation.