Add tools to the launcher
The LDAP SDK jar files include a launcher that can be used to invoke tools
using the command "java -jar {jarfile} {toolname} {args}".  This update
exposes the base64 and transform-ldif tools through the launch in both the
Standard Edition and Commercial Edition, and also exposes the
generate-totp-shared-secret tool in the Commercial Edition.