Fixed a multi-stage bind issue with message IDs.
The LDAP SDK formerly had a bug that would cause it to use the same message ID
for all stages of certain kinds of multi-stage SASL bind requests instead of
using a different message ID for each stage.  Although it's not likely to
cause problems in real-world use cases, it wasn't intentional and isn't ideal.