Fixed LDIF encoding for attributes without values.
When generating the LDIF representation of an entry, the LDAP SDK didn't
generate any output for attributes that didn't have any values.  Since
attribute values will generally only be omitted when searching with the
typesOnly flag set to true, this means that the output of the ldapsearch tool
would only include the entry's DN when invoked with the option to request only
attribute descriptions but not values.  For the purpose of generating the LDIF
representation, an attribute without any values will be treated like an
attribute with one zero-byte value.