Fix RDN string generation
The code used to generate the string representation of an RDN from its
components included some unnecessary escaping.  According to RFC 4514 section
2.4, the octothorpe character (#) only needs to be escaped if it appears at
the beginning of an attribute value, and spaces only need to be escaped if
they appear at the beginning or end of a value.  However, the code to generate
the string representation, including the normalized string representation,
would always escape the octothorpe character anywhere in the value, and would
escape all but the last space in a string of consecutive spaces.  This wasn't
technically wrong, since RFC 4514 does allow for more than the minimum amount
of escaping, but it has been observed to cause problems with certain clients,
so we no longer escape the octothorpe or space characters except as required
by the specification.