Use schema for in-memory directory server imports
Previously, there were cases in which the in-memory directory server calls to
the InMemoryDirectoryServer,importFromLDIF(boolean, String) and to
InMemoryDirectoryServer.addEntries(String...) methods would not use the schema
that had been configured for the server.  This could lead to adding entries
that violate the schema, and could lead to cases in which the appropriate
matching rules were not used for certain attributes.  This has been fixed.