Fixed an AuthenticationFailureReason conflict
The Commercial Edition of the LDAP SDK includes support for the
UnboundID-proprietary get password policy state issues request and response
controls.  These controls can be used to help identify the reason for a failed
authentication attempt, as well as any current or upcoming issues that might
affect the usability of the authenticated user's account.  The authentication
failure reason has an integer value, a name, and an optional message, and the
LDAP SDK inadvertently duplicated the integer value for two different failure
reasons (one indicating that the account isn't usable, and another indicating
that some unknown problem occurred).  At present, the Directory Server only
uses the "other" failure reason in a very specific context that is unlikely to
occur in the vast majority of environments, so I'm changing the integer value
associated with it to avoid this conflict.