Sign release builds of the LDAP SDK.
Updated the build process to generate a PGP signature for the LDAP SDK zip
file if the env.GPG_PASSPHRASE property is set (and it will be set for
official release builds of the LDAP SDK).  The build process already generated
signatures for components destined for inclusion in the Maven Central
Repository, but we will now also create a signature for the downloadable zip
file.