Fixed a connection pool defunct connection issue
Fixed an issue with the LDAP connection pool that could cause the pool to run
out of connections and refuse to create new ones if createIfNecessary is false
and replaceDefunctConnection is unable to create a new connection to replace
a connection that is no longer valid.

The pool did not properly record the failed attempt to create a new
connection, and therefore still assumed that a valid connection was checked
out.  This would cause the pool to operate with a maximum size that is one
less than it should be for each failed replaceDefunctConnection attempt, and
if it happened enough times, the pool would be out of connections and would
refuse to create any new ones.