Fixed a bug in parsing hex-encoded RDN values
The LDAP distinguished name specification in RFC 4514 allows an attribute
value included in the string representation of a DN or RDN to be provided as
a hex-encoded string, prefixed by the octothorpe character (#), specifying the
raw bytes that comprise that value.  In implementing the LDAP SDK's support
for parsing DNs and RDNs from their string values, this portion of the RFC was
misinterpreted.  The implementation assumed that the hex-encoded string was
comprised entirely of the bytes of the attribute value, but it is actually
supposed to be the complete set of bytes that comprise a BER element (the
type, length, and value), with the RDN attribute value being the value of that
BER element.  As a result, the LDAP SDK was both too lenient with the
hex-encoded values that it would accept (in that it would accept hex-encoded
values that do not represent valid BER elements), and the value that it parsed
incorrectly included the bytes that represented the BER type and length in
addition to the intended value.  This bug has been fixed.