Fix an interactive mode+properties file bug.
The command-line tool API supports an interactive mode that offers a
menu-driven interface that can walk a user through the process of providing
the arguments needed to run the tool.  Tools that support interactive mode
offer an "--interactive" argument that can force interactive mode, but they
can also be written so that they default to launching in interactive mode if
no arguments are provided on the command line.

Tools can also support reading the values for one or more arguments from a
properties file.  This can be convenient for cases in which you always run the
tool with a common set of arguments (for example, the arguments used to
connect and authenticate to a directory server).  If the tool supports a
properties file, you can use the "--propertiesFilePath" argument to specify
the path to the properties file to use, but it can also determine a default
properties file path from a system property
(com.unboundid.util.args.ArgumentParser.propertiesFilePath) or an environment
variable (UNBOUNDID_TOOL_PROPERTIES_FILE_PATH).

There was a bug that could arise in tools that support both interactive mode
and using a properties file, that default to running in interactive mode when
launched without any arguments, and when a default properties file is set via
system property or environment variable.  In such cases, the tool would
incorrectly start in interactive mode even if one or more argument values
should have been set from the properties file.  This bug has been fixed.