Include LDAP SDK version in exception messages
Updated a number of exception messages so that they will include the LDAP SDK
version number (in addition to the source revision that was already there) to
make it easier to identify the origin of a problem.

Also improved the LDAPExtendedOperationException.toString method so that it
will include the response OID and extended result name, if available.