Improve build-time repository info collection
For support purposes, the LDAP SDK build process has the ability to obtain
information about the repository from which the source code was retrieved.
Previously, only subversion repositories were supported (using the open source
SVNKit library).  This commit adds support for git repositories (using the
open source JGit library, which depends on the open source slf4j library).

The effects of this change are:

- A new com.unboundid.ldap.sdk.Version.REPOSITORY_TYPE constant is available.
  Its value will be one of "subversion", "git", or "{unknown}".

- A new com.unboundid.ldap.sdk.Version.REPOSITORY_URL constant is available.
  If the LDAP SDK source code was obtained from a subversion or git
  repository, then the appropriate type of URL should be used for that type of
  repository (e.g., https://svn.code.sf.net/p/ldap-sdk/code or
  git@github.com:pingidentity/ldapsdk.git).  For any other type of repository,
  it will be a file URL pointing that references the path to the source on the
  build machine filesystem (e.g., "file:///workspaces/ldap-sdk").

- A new com.unboundid.ldap.sdk.Version.REVISION_ID constant is available.  Its
  value will be a repository-appropriate identifier for the source repository.
  For a subversion repository, this will be the string representation of the
  revision number.  For a git repository, this will be the SHA-1 digest of the
  most recent commit.  For any other type of repository, the value will be
  "{unknown}".

- The com.unboundid.ldap.sdk.Version.REVISION_NUMBER constant is now
  deprecated in favor of the REVSION_ID constant.  It will continue to be
  maintained, but since some types of repositories (like git) do not offer
  numeric revision identifiers, it is not always useful.  If the LDAP SDK was
  checked out from a subversion repository then the value should be accurate,
  but the value will be -1 in all other circumstances.

- All references to the Version.REVISION_NUMBER constant in the LDAP SDK have
  been updated to us Version.REVISION_ID instead.

Note that while this commit includes code that depends on the JGit and slf4j
libraries, those libraries are build-time dependencies only (and they are
checked into the repository, so there is no need to go looking for them).
These libraries are not required in any way when using the LDAP SDK, and the
LDAP SDK's only runtime dependency is Java SE 7 or 8.