Fixed in-memory directory server subentry issues
Updated the in-memory directory server to fix a couple of issues with the way
that it handled LDAP subentries.  The server would incorrectly return
non-subentries when processing search requests that included the subentries
request control, and the server would incorrectly omit subentries when
processing search requests with a non-baseObject scope in which the base entry
was itself a subentry.