Fixed an OSGi problem in the jar file manifest.
Updated the LDAP SDK build process to correct the value of the
Bundle-RequiredExecutionEnvironment property in the jar file manifest, which
is used by OSGi.  When the LDAP SDK supported Java 1.5 or later, the value
of this property was "J2SE-1.5", and when we updated it to require Java 7 or
later, the property was updated to have a value of "J2SE-1.7".  However, the
format of the value changed with Java 1.6, and the "J2SE" has been replaced
with "JavaSE", so the correct property value should be "JavaSE-1.7".  This
has now been corrected.