Password encoding support for the in-memory server
The in-memory directory server has been updated to provide support for working
with encoded passwords.  In the past, the in-memory directory server has
always stored passwords in the clear, with no alteration of any kind.  That is
still the default, but it is now possible to configure the server to
automatically encode new clear-text passwords using a specified mechanism, and
to support existing pre-encoded passwords using a provided set of mechanisms.

This functionality is exposed through the InMemoryDirectoryServerConfig class,
which now offers a setPasswordEncoders method that allows you to set an
optional primary password encoder (that will be used to encode any clear-text
passwords provided for storage in the server), and an optional collection of
secondary encoders (that will be used to interact with pre-encoded passwords,
but that will not be used to encode new passwords).

It is also possible to configure the set of password attributes that will be
used by the in-memory directory server.  Previously, this was hard-coded to
be userPassword, and that is still the default configuration, but you can now
use the setPasswordAttributes method to specify which attribute(s) are
intended to hold passwords.  If the server has been configured with a primary
password encoder, then any clear-text values provided for these password
attributes in an LDIF import, LDAP add request, LDAP modify request, or
password modify extended request, will be automatically encoded.  If the
server has been configured with primary or secondary password encoders, then
it can make use of those encoders to compare clear-text bind passwords against
the encoded values stored in the server.

The server provides pluggable support for password encoders via the extensible
com.unboundid.ldap.listener.InMemoryPasswordEncoder class.  Each encoder must
define a unique prefix (for example, you might use "{SHA}" for an encoder that
uses an unsalted SHA-1 digest), and it can also use an output formatter that
transforms the output in some way.  The server includes password encoder
implementations that leave passwords in the clear or that use a salted or an
unsalted message digest.  It also includes output formatter implementations
represent raw bytes in base64 or hexadecimal.