Update ldapsearch to make the base DN optional
Previously, the ldapsearch tool required exactly one of the --baseDN and
--ldapURLFile arguments to be present.  This could result in awkward behavior
when trying to run the tool in interactive mode with an empty base DN because
you would have to explicitly select the base DN item from the menu and then
press ENTER when asked for a value.

With this change, the --baseDN argument is made optional.  You still can't
provide it in conjunction with the --ldapURLFile argument, but if neither of
those arguments is given, then the ldapsearch tool will now use a default of
the null base DN (the empty string).