Add another ResultCode.valueOf method
The ResultCode.valueOf method allows you to get a ResultCode object with a
specified integer value.  Previously, there were two versions of that method:
one that just takes an integer argument (the numeric value for the result code
to retrieve), and the other took an integer and a string (the human-readable
name to use for the result code if it doesn't already exist).  Both of these
methods would always create a new ResultCode object if there wasn't already
one with the provided integer value (and cache it to be returned for
subsequent calls with that value).  The new version of the method allows you
to indicate whether to create a new ResultCode object if there isn't already
one with the requested integer value or to return null to indicate that it
doesn't already exist.