Added a JVM-default turst manager
Added a new trust manager that is configured to use the JVM's default trusted
issuer store, which probably exists in the ${JAVA_HOME}/lib/security/cacerts or
the ${JAVA_HOME}/jre/lib/security/cacerts file.  The command-line tool
interfaces have been updated to augment the places where the prompt trust
manager would have been used in the past to combine it with the new
JVM-default trust manager so that you will no longer be prompted for
certificates that are signed by an issuer that the JVM considers trusted.